# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "React lab API",
    "description": "API on Node.js for ITRex React.js Lab",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "Tsimafei Kinevich",
      "url": "https://itrexgroup.com/",
      "email": "tsimafei.kinevich@itrexgroup.com"
    }
  },
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/api/appointments": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Create new appointment",
        "security": [{
          "Bearer": []
        }],
        "parameters": [{
          "name": "Request body",
          "in": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "date": {
                "example": "2021-11-23 13:00",
                "type": "string"
              },
              "reason": {
                "example": "stomach-ache",
                "type": "string"
              },
              "note": {
                "example": "stomach-ache",
                "type": "string"
              },
              "doctorID": {
                "example": "882240b7-514a-48c3-a67f-6acfb1d35030",
                "type": "string"
              }
            }
          }
        }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Appointment"
            }
          }
        }
      }
    },
    "/api/appointments/doctor/me": {
      "get": {
        "tags": ["Appointments"],
        "summary": "Get all appointments for doctors",
        "security": [{
          "Bearer": []
        }],
        "parameters": [{
          "name": "offset",
          "in": "query",
          "description": "Offset for pagination",
          "required": true,
          "type": "number"
        },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit for pagination",
            "required": true,
            "type": "number"
          },
          {
            "name": "name",
            "in": "query",
            "description": "Doctor name filter",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort by",
            "required": false,
            "type": "string",
            "enum": ["firstNameSort", "lastNameSort", "dateSort"]
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort variant",
            "required": false,
            "type": "string",
            "enum": ["Asc", "Desc"]
          },
          {
            "name": "dateStatus",
            "in": "query",
            "description": "appointment date status",
            "required": false,
            "type": "string",
            "enum": ["Upcoming", "Outdate"]
          }
        ],
        "responses": {
          "200": {
            "description": "Doctors",
            "schema": {
              "type": "object",
              "properties": {
                "appointments": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/AppointmentForDoctor"
                  }
                },
                "total": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/{id}": {
      "parameters": [{
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Appointment ID",
        "type": "string"
      }],
      "delete": {
        "tags": ["Appointments"],
        "summary": "Delete an appointment",
        "security": [{
          "Bearer": []
        }],
        "responses": {
          "200": {
            "description": "Deleted",
            "schema": {
              "description": "Deleted appointment ID",
              "type": "string"
            }
          }
        }
      },
      "patch": {
        "tags": ["Appointments"],
        "summary": "Update a appointment status",
        "security": [{
          "Bearer": []
        }],
        "parameters": [{
          "name": "Request body",
          "in": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "status": {
                "example": "Waiting for confirmation",
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "Updated appointment ID",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/patient/me": {
      "get": {
        "tags": ["Appointments"],
        "summary": "Get all appointments for patients",
        "security": [{
          "Bearer": []
        }],
        "parameters": [{
          "name": "offset",
          "in": "query",
          "description": "Offset for pagination",
          "required": true,
          "type": "number"
        },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit for pagination",
            "required": true,
            "type": "number"
          },
          {
            "name": "name",
            "in": "query",
            "description": "Doctor name filter",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort by",
            "required": false,
            "type": "string",
            "enum": ["firstNameSort", "lastNameSort", "dateSort"]
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort variant",
            "required": false,
            "type": "string",
            "enum": ["Asc", "Desc"]
          },
          {
            "name": "dateStatus",
            "in": "query",
            "description": "appointment date status",
            "required": false,
            "type": "string",
            "enum": ["Upcoming", "Outdate"]
          }
        ],
        "responses": {
          "200": {
            "description": "Doctors",
            "schema": {
              "type": "object",
              "properties": {
                "appointments": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/AppointmentForPatient"
                  }
                },
                "total": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/time/free": {
      "get": {
        "tags": ["Appointments"],
        "summary": "Get a free time for visit",
        "parameters": [{
          "name": "date",
          "in": "query",
          "description": "date for visit",
          "required": true,
          "type": "string"
        },
          {
            "name": "doctorID",
            "in": "query",
            "description": "Doctor ID",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Doctor data",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "user login API",
        "parameters": [{
          "name": "Request body",
          "in": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "login": {
                "example": "dima@mail.ru",
                "maxLength": 30,
                "minLength": 2,
                "type": "string"
              },
              "password": {
                "example": "dima@mail.ru",
                "maxLength": 30,
                "minLength": 2,
                "type": "string"
              }
            }
          }
        }
        ],
        "responses": {
          "200": {
            "description": "User data",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/AuthData"
            }
          }
        }
      }
    },
    "/api/auth/password": {
      "patch": {
        "tags": ["Auth"],
        "summary": "User change password API",
        "security": [{
          "Bearer": []
        }],
        "parameters": [{
          "name": "Request body",
          "in": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "oldPassword": {
                "example": "dima@mail.ru",
                "maxLength": 30,
                "minLength": 2,
                "type": "string"
              },
              "newPassword": {
                "example": "DimaPassword",
                "maxLength": 30,
                "minLength": 2,
                "type": "string"
              }
            }
          }
        }
        ],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/auth/registration": {
      "post": {
        "tags": ["Auth"],
        "summary": "user registration API",
        "parameters": [
          {
            "name": "Request body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "login": {
                  "example": "dima@mail.ru",
                  "maxLength": 30,
                  "minLength": 2,
                  "type": "string"
                },
                "password": {
                  "example": "dima@mail.ru",
                  "maxLength": 30,
                  "minLength": 2,
                  "type": "string"
                },
                "firstName": {
                  "example": "Dima",
                  "maxLength": 30,
                  "minLength": 2,
                  "type": "string"
                },
                "lastName": {
                  "example": "Mitrichenko",
                  "maxLength": 30,
                  "minLength": 2,
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/AuthData"
            }
          }
        }
      }
    },
    "/api/auth/profile": {
      "post": {
        "tags": ["Auth"],
        "summary": "check user access",
        "security": [{
          "Bearer": []
        }],
        "responses": {
          "200": {
            "description": "Access status",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "first_name": {
                  "type": "string"
                },
                "last_name": {
                  "type": "string"
                },
                "login": {
                  "type": "string",
                  "uniqueItems": true
                },
                "photo": {
                  "type": "string",
                  "description": "Path to photo"
                },
                "role_name": {
                  "type": "string",
                  "description": "User role"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/token/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh JWT token ",
        "security": [{
          "Bearer": []
        }],
        "responses": {
          "200": {
            "description": "Access status",
            "schema": {
              "type": "object",
              "properties": {
                "access_token" : {
                  "type": "string"
                },
                "refresh_token" : {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AppointmentForDoctor": {
      "properties": {
        "id": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "reason": {
          "type": "string"
        },
        "note": {
          "type": "string",
          "description": "Note for visit"
        },
        "photo": {
          "type": "string",
          "description": "Path to photo"
        },
        "patient_id": {
          "type": "string"
        },
        "doctor_id": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "description": "Appointment status name"
        }
      }
    },
    "AppointmentForPatient": {
      "properties": {
        "id": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "reason": {
          "type": "string"
        },
        "note": {
          "type": "string",
          "description": "Note for visit"
        },
        "photo": {
          "type": "string",
          "description": "Path to photo"
        },
        "patient_id": {
          "type": "string"
        },
        "doctor_id": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "description": "Appointment status name"
        },
        "specialization_name": {
          "type": "string",
          "description": "Doctor specialization name"
        }
      }
    },
...
    "AuthData": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        },
        "refresh_token": {
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "photo": {
          "type": "string"
        },
        "role_id": {
          "type": "string"
        }
      }
    },
    "Appointment": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "patient_id": {
          "type": "string"
        },
        "doctor_id": {
          "type": "string"
        },
        "visit_date": {
          "type": "string"
        },
        "reason": {
          "type": "string"
        },
        "note": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}